frontend/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "A.G. Freitas", email = "agf.adriano@gmail.com"},
]
dependencies = [
    "gradio>=3.45.2",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11,<4"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "black>=23.9.1",
    "flake8>=6.1.0",
    "isort>=5.12.0",
]

[tool.black]
line-length = 79

[tool.pdm.scripts]
chatbot = "python src/app/main.py"
